requirements.txt
Flask
Flask-Bcrypt
Flask-Login
Flask-SQLAlchemy
FPDF
pytesseract
Pillow
